Add better support for mesh generation

+++ mesh.cpp
@@ -1,4 +1,4 @@
-#include <igl/opengl/glfw/Viewer.h>
+#include "mesh.h"
 
 /*
  * Generate a mesh with a square linkage pattern, with dimensions N x M (where N = number of rows of squares, M = number
@@ -6,7 +6,7 @@
  *
  * Return the vertex positions and face vertex indices in the matrices V and F, respectively.
  */
-void generateSquarePatternMesh(size_t N, size_t M, Eigen::MatrixXd& V, Eigen::MatrixXi& F) {}
+void generateSquarePatternMesh(int N, int M, Eigen::MatrixXd& V, Eigen::MatrixXi& F) {}
 
 
 /* TODO: Feel free to generate other linkage patterns, such as rectangle, parallelogram, etc.! */
